+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.47.0
+
+- `SQLAlchemyDataLayer` now works with both, legacy `orm.Query` and new `sa.Select`
+  objects transparently
+- fix: unsafe and undefined behavior in SQLAlchemy filtering
+- small fixes in `SQLAlchemyDataLayer` handling of relationships
+
 ## 0.46.0
 
 - fix: using string filters on enum fields didn't work with SQLAlchemy